fix(achievements): replace bulk sync with per-game GetPlayerAchievements
The bulk GetTopAchievementsForGames endpoint cannot provide the data the
app needs. Verified against the raw response for Portal 2: achievements
are identified by statid+bit and the only text field is a *localized*
`name` ("Pturretdactyl", "Drop Box", ...). There is no apiname in the
response, so the rows we wrote into user_achievements could never join
against game_achievements — the library card rendered 0/51 for any game
that only went through the bulk path. The endpoint is also lossy: even
at max_achievements=10000 it returns only top unlocks, which is why
sync reported ~3680 of ~18000 real unlocks for this account.

Swap syncAchievementsForStats to fan out per-game GetPlayerAchievements
requests with a concurrency cap of 5, piggybacking on the existing
persistAchievements + ensureSchema code paths. Drops persistBulkGameStats
and the getTopAchievementsForGames wrapper entirely.
