+++ lib/steam-api.ts
@@ -143,70 +143,6 @@ export async function getPlayerAchievements(steamId: string, appId: number): Pro
   }
 }
 
-export interface BulkTopAchievement {
-  statid?: number
-  bit?: number
-  name?: string
-  desc?: string
-  icon?: string
-  icon_gray?: string
-  hidden?: boolean
-  player_percent_unlocked?: string
-}
-
-export interface BulkGameAchievements {
-  appid: number
-  total_achievements: number
-  achievements?: BulkTopAchievement[]
-}
-
-/**
- * Fetches top achievement data for multiple games in a single request using
- * the undocumented IPlayerService/GetTopAchievementsForGames endpoint.
- *
- * The response only includes **unlocked** achievements for each game; locked
- * achievements are implied by `total_achievements - achievements.length`.
- * Games the user hasn't unlocked anything in are returned with just
- * `total_achievements` and no `achievements` array.
- *
- * Returns an empty array if the request fails (the caller can fall back to
- * per-game syncs for the affected batch).
- */
-export async function getTopAchievementsForGames(
-  steamId: string,
-  appIds: number[],
-  maxAchievements = 10_000,
-): Promise<BulkGameAchievements[]> {
-  if (appIds.length === 0) return []
-
-  const apiKey = process.env.STEAM_API_KEY
-  if (!apiKey) {
-    throw new SteamAPIError("STEAM_API_KEY is not configured", 500)
-  }
-
-  const url = new URL(`${STEAM_API_BASE}/IPlayerService/GetTopAchievementsForGames/v1/`)
-  url.searchParams.set("key", apiKey)
-  url.searchParams.set("format", "json")
-  url.searchParams.set("steamid", steamId)
-  url.searchParams.set("language", "es")
-  url.searchParams.set("max_achievements", String(maxAchievements))
-  appIds.forEach((appId, index) => {
-    url.searchParams.set(`appids[${index}]`, String(appId))
-  })
-
-  try {
-    const response = await fetch(url.toString(), { cache: "no-store" })
-    if (!response.ok) {
-      throw new SteamAPIError(`Steam API request failed: ${response.status}`, response.status)
-    }
-    const data = (await response.json()) as { response?: { games?: BulkGameAchievements[] } }
-    return data.response?.games ?? []
-  } catch (error) {
-    console.error("Error fetching bulk achievements for batch of", appIds.length, "games:", error)
-    return []
-  }
-}
-
 /** Fetches the game schema (achievement and stat definitions) from the Steam API. */
 export async function getGameSchema(appId: number): Promise<unknown> {
   try {
